Update and rename nzbdrone to sonarr
Note that not all filenames are changed at first. If you update to this file, make sure to change the name of the folder the application is in to sonarr, but do NOT change NzbDrone.exe to sonarr.exe. I can only assume the .exe-file will be changed at some point, but that's for upstream.

This also changes both username and groupname, not to mention log location, data_dir and so forth. As such, I am not committing this to master yet.